Add store_people and read_people
