+++ JSONdata/books.json
@@ -6,5 +6,9 @@
   {
     "title" : "Think like a man",
     "author" : "Bradley Logan"
+  },
+  {
+    "title" : "The smart money woman",
+    "author" : "Arese Ugwu"
   }
 ]
