+++ JSONdata/people.json
@@ -1,0 +1,37 @@
+[
+  {
+    "id" : 505,
+    "type" : "Student",
+    "age" : 34,
+    "name" : "Rex",
+    "parent_permission" : true
+  },
+  {
+    "id" : 47,
+    "type" : "Student",
+    "age" : 45,
+    "name" : "Milani Mac",
+    "parent_permission" : true
+  },
+  {
+    "id" : 80,
+    "type" : "Student",
+    "age" : 19,
+    "name" : "Adam Eve",
+    "parent_permission" : true
+  },
+  {
+    "id" : 644,
+    "type" : "Student",
+    "age" : 23,
+    "name" : "Abigail",
+    "parent_permission" : true
+  },
+  {
+    "id" : 995,
+    "type" : "Teacher",
+    "age" : 34,
+    "name" : "Mrs Anderson",
+    "specialization" : "Basic Tech"
+  }
+]
