--- data/people.json
@@ -1,23 +1,0 @@
-[
-  {
-    "id" : 405,
-    "type" : "Student",
-    "age" : 34,
-    "name" : "Rex",
-    "parent_permission" : true
-  },
-  {
-    "id" : 976,
-    "type" : "Teacher",
-    "age" : 45,
-    "name" : "Milani Mac",
-    "parent_permission" : true
-  },
-  {
-    "id" : 29,
-    "type" : "Student",
-    "age" : 19,
-    "name" : "Adam Eve",
-    "parent_permission" : true
-  }
-]
