+++ teacher.rb
@@ -1,6 +1,8 @@
 require_relative 'person'
 
 class Teacher < Person
+  attr_accessor :specialization
+  
   def initialize(age, name, specialization)
     super(age, name, parent_permission: true)
     @specialization = specialization
